+++ src/single_range.h
@@ -33,8 +33,8 @@ namespace MultiArrayTools
 	virtual SingleIndex& operator++() override;
 	virtual SingleIndex& operator--() override;
 
-	virtual size_t pp(std::shared_ptr<IndexBase>& idxPtr) override;
-	virtual size_t mm(std::shared_ptr<IndexBase>& idxPtr) override;
+	virtual int pp(std::shared_ptr<IndexBase>& idxPtr) override;
+	virtual int mm(std::shared_ptr<IndexBase>& idxPtr) override;
 	
 	virtual U meta() const override;
 	virtual SingleIndex& at(const U& metaPos) override;
